+++ CHANGELOG.md
@@ -4,7 +4,16 @@ Changes to Webiblioteca Project will be documented in this file.
 ## Unreleased
 - Edit book module;
 - Options menu;
+
+## Beta 0.5.0 | 2022-8-29
+#### Added
 - About menu;
+
+#### Fixed
+- Minor error fixes;
+
+#### Changed
+- Default exemple;
 
 ## Beta 0.4.0 | 2022-8-28
 #### Added